# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PYQDeck API Explorer",
    "version": "1.0.0",
    "description": "\nWelcome to the official **PYQDeck API Documentation**.\n\nThis API powers the PYQDeck platform, enabling users to access, bookmark, and study Previous Year Questions (PYQs) efficiently.\n\n### 🔑 Authentication\nMost endpoints require a valid Clerk JWT token. Include it in the header:\n`Authorization: Bearer <your_clerk_token>`\n\n### 📊 Rate Limiting\n- **Standard API**: 100 requests per 15 minutes.\n- **Webhooks**: 50 requests per 15 minutes.\n",
    "contact": {
      "name": "PYQDeck Support",
      "url": "http://localhost:3000/support",
      "email": "noreply@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request was invalid or could not be understood.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication is required and has failed or has not yet been provided.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "The requested resource could not be found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string",
            "example": "Detailed error message here"
          },
          "code": {
            "type": "integer",
            "example": 400
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "code": {
            "type": "integer",
            "example": 200
          }
        }
      },
...
      "PermissionGrant": {
        "type": "object",
        "required": [
          "userId",
          "capabilities",
          "scopeLevel"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "65a12345b67890cdef444444"
          },
          "userId": {
            "type": "string",
            "description": "Reference to User this grant belongs to",
            "example": "65b98765a43210fedcba9876"
          },
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "content:create",
                "content:edit",
                "content:moderate",
                "content:delete"
              ]
            },
            "example": [
              "content:create",
              "content:moderate"
            ]
          },
          "scopeLevel": {
            "type": "string",
            "enum": [
              "global",
              "university",
              "branch",
              "semester",
              "subjectOffering"
            ],
            "example": "branch"
          },
          "scopeId": {
            "type": "string",
            "nullable": true,
            "description": "Required unless scopeLevel is \"global\" -- points at the University/Branch/Semester/SubjectOffering this grant is scoped to",
            "example": "65a12345b67890cdef111111"
          },
          "label": {
            "type": "string",
            "description": "Human-readable name set at creation time, e.g. \"Content Moderator -- Computer Engineering\"",
            "example": "Content Moderator -- Computer Engineering"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          },
          "grantedBy": {
            "type": "string",
            "description": "Reference to the admin User who created this grant"
          },
          "revokedBy": {
            "type": "string",
            "nullable": true
          },
          "revokedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "notes": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "User": {
        "type": "object",
        "required": [
          "clerkId",
          "name",
          "email"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The auto-generated id of the user",
            "example": "65b98765a43210fedcba9876"
          },
          "clerkId": {
            "type": "string",
            "description": "The unique ID from Clerk authentication",
            "example": "user_2N9Wv7N0u8M6X1Y2Z3A4B5C6"
          },
          "name": {
            "type": "string",
            "description": "The name of the user",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "The email of the user",
            "example": "john.doe@example.com"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "description": "The profile image URL from Clerk",
            "example": "https://img.clerk.com/..."
          },
          "universityId": {
            "type": "string",
            "nullable": true,
            "description": "Reference to University",
            "example": "60d0fe4f5311236168a109ca"
          },
          "branchId": {
            "type": "string",
            "nullable": true,
            "description": "Reference to Branch",
            "example": "65a12345b67890cdef111111"
          },
          "semesterId": {
            "type": "string",
            "nullable": true,
            "description": "Reference to Semester",
            "example": "65a12345b67890cdef112233"
          },
          "role": {
            "type": "string",
            "enum": [
              "normal",
              "admin",
              "editor"
            ],
            "default": "normal",
            "description": "The role of the user",
            "example": "normal"
          },
          "isActive": {
            "type": "boolean",
            "default": true,
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "getCurrentUser",
        "tags": [
          "Users"
        ],
        "summary": "Get my profile and activity stats",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User record and bookmark counts",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "user": {
                              "$ref": "#/components/schemas/User"
                            },
                            "stats": {
                              "type": "object",
                              "properties": {
                                "bookmarks": {
                                  "type": "integer"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users": {
      "get": {
        "operationId": "listUsers",
        "tags": [
          "Users"
        ],
        "summary": "List all users (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "editor",
                "normal"
              ]
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "email",
                "role",
                "createdAt"
              ],
              "default": "createdAt"
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users with pagination",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/User"
                              }
                            },
                            "total": {
                              "type": "integer"
                            },
                            "page": {
                              "type": "integer"
                            },
                            "limit": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/users/{clerkId}": {
      "get": {
        "operationId": "getUserById",
        "tags": [
          "Users"
        ],
        "summary": "Get a user by clerkId with activity stats (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clerkId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User record and activity stats",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "user": {
                              "$ref": "#/components/schemas/User"
                            },
                            "stats": {
                              "type": "object",
                              "properties": {
                                "bookmarksCount": {
                                  "type": "integer"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "operationId": "updateUser",
        "tags": [
          "Users"
        ],
        "summary": "Update a user (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clerkId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "editor",
                      "normal"
                    ]
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "user": {
                              "$ref": "#/components/schemas/User"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/users/{clerkId}/grants": {
      "get": {
        "operationId": "listUserGrants",
        "tags": [
          "Users"
        ],
        "summary": "List a user's scoped permission grants (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clerkId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "includeRevoked",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of permission grants",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/PermissionGrant"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "operationId": "createUserGrant",
        "tags": [
          "Users"
        ],
        "summary": "Grant a user a scoped permission (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clerkId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "capabilities",
                  "scopeLevel"
                ],
                "properties": {
                  "capabilities": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "content:create",
                        "content:edit",
                        "content:moderate",
                        "content:delete"
                      ]
                    }
                  },
                  "scopeLevel": {
                    "type": "string",
                    "enum": [
                      "global",
                      "university",
                      "branch",
                      "semester",
                      "subjectOffering"
                    ]
                  },
                  "scopeId": {
                    "type": "string",
                    "nullable": true
                  },
                  "label": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Grant created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "grant": {
                              "$ref": "#/components/schemas/PermissionGrant"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/users/{clerkId}/grants/{grantId}": {
      "delete": {
        "operationId": "revokeUserGrant",
        "tags": [
          "Users"
        ],
        "summary": "Revoke a user's scoped permission grant (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clerkId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "grantId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Grant revoked",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "grant": {
                              "$ref": "#/components/schemas/PermissionGrant"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}